+++ lib/oneview-sdk/resource/api300/c7000/interconnect.rb
@@ -14,7 +14,19 @@ require_relative '../../api200/interconnect'
 module OneviewSDK
   module API300
     module C7000
+      # Interconnect resource implementation on API300 C7000
       class Interconnect < OneviewSDK::API200::Interconnect
+
+        # Create a resource object, associate it with a client, and set its properties.
+        # @param [OneviewSDK::Client] client The client object for the OneView appliance
+        # @param [Hash] params The options for this resource (key-value pairs)
+        # @param [Integer] api_ver The api version to use when interracting with this resource.
+        def initialize(client, params = {}, api_ver = nil)
+          @data ||= {}
+          # Default values:
+          @data['type'] ||= 'InterconnectV300'
+          super
+        end
       end
     end
   end