+++ spec/integration/resource/api200/uplink_set/delete_spec.rb
@@ -6,9 +6,10 @@ RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
 
   describe '#delete' do
     it 'delete resource' do
-      item = OneviewSDK::UplinkSet.new($client, name: UPLINK_SET_NAME)
+      item = klass.new($client, name: UPLINK_SET4_NAME)
       item.retrieve!
       expect { item.delete }.not_to raise_error
+      expect(item.retrieve!).to eq(false)
     end
   end
 end