+++ spec/integration/resource/api300/c7000/logical_interconnect_group/delete_spec.rb
@@ -11,6 +11,8 @@ RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
     @item_2.retrieve!
     @item_3 = klass.new($client_300, name: LOG_INT_GROUP3_NAME)
     @item_3.retrieve!
+    @item_4 = klass.new($client_300, name: LOG_INT_GROUP4_NAME)
+    @item_4.retrieve!
   end
 
   describe '#delete' do
@@ -18,6 +20,7 @@ RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
       expect { @item.delete }.to_not raise_error
       expect { @item_2.delete }.to_not raise_error
       expect { @item_3.delete }.to_not raise_error
+      expect { @item_4.delete }.to_not raise_error
     end
   end
 end